Non-function-function changing
Moved Text, CDATA, etc. to (new) XML class.
