+++ src/main/scala/com/phasmidsoftware/core/Utilities.scala
@@ -1,8 +1,6 @@
 package com.phasmidsoftware.core
 
 import cats.effect.IO
-import com.phasmidsoftware.kmldoc.Mergeable
-import com.phasmidsoftware.xml.{CDATA, Extractor, Extractors}
 import scala.collection.mutable
 import scala.util.matching.Regex
 import scala.xml.{Elem, Node, NodeSeq}
@@ -53,43 +51,6 @@ object Utilities {
   def show(node: Node)(f: String => Unit = println): IO[Unit] = IO(f(renderNode(node)))
 }
 
-case class Text($: CharSequence) extends Mergeable[Text] {
-  /**
-   * Method to merge this and another Text element.
-   *
-   * TODO: we should add a delimiter
-   * TODO: we should merge CDATA elements
-   *
-   * @param t the object to be merged with this.
-   * @return the merged value of T.
-   */
-  def merge(t: Text): Option[Text] = Some(Text($.toString ++ t.$.toString))
-
-  override def equals(obj: Any): Boolean = obj match {
-    case text: Text => $ == text.$
-    case _ => false
-  }
-}
-
-object Text extends Extractors {
-
-  def namesMatch(name: Text, name1: String): Boolean = name.$ match {
-    case c: CDATA => c.toString == name1
-    case x: CharSequence => x.toString == name1
-    case _ => false
-  }
-
-  /**
-   * Text extractor.
-   */
-  implicit val extractorText: Extractor[Text] = extractor10(apply)
-
-  /**
-   * Optional text extractor.
-   */
-  implicit val extractorOptionalText: Extractor[Option[Text]] = extractorOption[Text]
-
-}
 
 class MappableRegex(r: String, f: List[String] => List[String]) extends Regex(r) {
   override def unapplySeq(s: CharSequence): Option[List[String]] =
